move to browser tab  utils getMaskedUrl and isEnsUrl function
