+++ app/components/Views/BrowserTab/utils.ts
@@ -1,5 +1,7 @@
+import AppConstants from '../../../core/AppConstants';
 import { prefixUrlWithProtocol } from '../../../util/browser';
 import URLParse from 'url-parse';
+import { SessionENSNames } from './types';
 
 /**
  * Validates url for browser
@@ -37,4 +39,59 @@ export const processUrlForBrowser = (urlString: string): string => {
     return `https://www.google.com/search?q=${encodeURIComponent(urlString)}`;
   }
   return urlWithProtocol;
+};
+
+/**
+ * Checks if it is a ENS website
+ */
+export const isENSUrl = (urlToCheck: string, ensIgnoreList: string[]) => {
+  const { hostname } = new URLParse(urlToCheck);
+  const tld = hostname.split('.').pop();
+  if (
+    tld &&
+    AppConstants.supportedTLDs.indexOf(
+      tld.toLowerCase() as 'eth' | 'xyz' | 'test',
+    ) !== -1
+  ) {
+    // Make sure it's not in the ignore list
+    if (ensIgnoreList.indexOf(hostname) === -1) {
+      return true;
+    }
+  }
+  return false;
+};
+
+/**
+ * Gets the url to be displayed to the user
+ * For example, if it's ens then show [site].eth instead of ipfs url
+ */
+export const getMaskedUrl = (
+  urlToMask: string,
+  sessionENSNames: SessionENSNames,
+) => {
+  if (!urlToMask) return urlToMask;
+  let replace = null;
+  if (urlToMask.startsWith(AppConstants.IPFS_DEFAULT_GATEWAY_URL)) {
+    replace = (key: string) =>
+      `${AppConstants.IPFS_DEFAULT_GATEWAY_URL}${sessionENSNames[key].hash}/`;
+  } else if (urlToMask.startsWith(AppConstants.IPNS_DEFAULT_GATEWAY_URL)) {
+    replace = (key: string) =>
+      `${AppConstants.IPNS_DEFAULT_GATEWAY_URL}${sessionENSNames[key].hostname}/`;
+  } else if (urlToMask.startsWith(AppConstants.SWARM_DEFAULT_GATEWAY_URL)) {
+    replace = (key: string) =>
+      `${AppConstants.SWARM_DEFAULT_GATEWAY_URL}${sessionENSNames[key].hash}/`; //TODO: This was SWARM_GATEWAY_URL before, it was broken, understand what it does
+  }
+
+  if (replace) {
+    const key = Object.keys(sessionENSNames).find((ens) =>
+      urlToMask.startsWith(ens),
+    );
+    if (key) {
+      urlToMask = urlToMask.replace(
+        replace(key),
+        `https://${sessionENSNames[key].hostname}/`,
+      );
+    }
+  }
+  return urlToMask;
 };
